Add tbc/wrath versions for CMSG_STANDSTATECHANGE
