+++ wow_world_messages/src/world/shared/mod.rs
@@ -121,6 +121,7 @@ pub mod cmsg_set_target_obsolete_vanilla_tbc_wrath;
 pub mod cmsg_set_watched_faction_vanilla_tbc_wrath;
 pub mod cmsg_setsheathed_vanilla_tbc_wrath;
 pub mod cmsg_split_item_vanilla_tbc;
+pub mod cmsg_standstatechange_vanilla_tbc_wrath;
 pub mod cmsg_swap_inv_item_vanilla_tbc_wrath;
 pub mod cmsg_swap_item_vanilla_tbc_wrath;
 pub mod cmsg_time_sync_resp_tbc_wrath;
