+++ wow_world_messages/src/world/tbc/mod.rs
@@ -300,6 +300,8 @@ pub(crate) mod cmsg_setsheathed;
 pub use cmsg_setsheathed::*;
 pub(crate) mod cmsg_split_item;
 pub use cmsg_split_item::*;
+pub(crate) mod cmsg_standstatechange;
+pub use cmsg_standstatechange::*;
 pub(crate) mod cmsg_swap_inv_item;
 pub use cmsg_swap_inv_item::*;
 pub(crate) mod cmsg_swap_item;
