+++ wow_world_messages/src/world/wrath/cmsg_standstatechange.rs
@@ -1,0 +1,2 @@
+pub use crate::shared::cmsg_standstatechange_vanilla_tbc_wrath::CMSG_STANDSTATECHANGE;
+
