+++ wowm_language/src/docs/cmsg_standstatechange.md
@@ -1,6 +1,6 @@
 # CMSG_STANDSTATECHANGE
 
-## Client Version 1.12
+## Client Version 1.12, Client Version 2, Client Version 3
 
 ### Wowm Representation
